Refactor kde bandwidth estimation methods
